service sistem

+++ core/apps/accounts/models.py
@@ -39,20 +39,15 @@ class UserProfile(models.Model):
         default=False
     )
 
-    last_latitude = models.DecimalField(
-        max_digits=10,
-        decimal_places=8,
+    last_latitude = models.FloatField(
         null=True,
         blank=True
     )
 
-    last_longitude = models.DecimalField(
-        max_digits=11,
-        decimal_places=8,
+    last_longitude = models.FloatField(
         null=True,
         blank=True
     )
-
     created_at = models.DateTimeField(
         auto_now_add=True
     )
