+++ core/apps/master/Desa/models.py
@@ -11,16 +11,12 @@ class Desa(models.Model):
 
     nama = models.CharField(max_length=255)
 
-    latitude = models.DecimalField(
-        max_digits=10,
-        decimal_places=8,
+    latitude = models.FloatField(
         null=True,
         blank=True
     )
 
-    longitude = models.DecimalField(
-        max_digits=11,
-        decimal_places=8,
+    longitude = models.FloatField(
         null=True,
         blank=True
     )
